Fix DeadMethodRemoval when superclasses could not be loaded
This change effectively disables this transformer for classes
with supertypes (classes or interfaces) that are not either included
with the obfuscated JAR, or available to load (i.e. on the system
classpath), as without access to these types we cannot know if the
parent method is abstract (and so removing the method would break
compilation, even if it is not used).

A warning will be logged when this occurs.